instagram_scraper/requirements.txt
playwright==1.44.0
playwright-stealth==1.0.6
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
